+++ conf/resources/examples/demos/02_Lists.scala
@@ -4,22 +4,22 @@ import leon.annotation._
 
 object List {
   sealed abstract class List
-  case class Cons(head: Int, tail: List) extends List
+  case class Cons(head: BigInt, tail: List) extends List
   case object Nil extends List
 
-  def size(l: List) : Int = (l match {
+  def size(l: List) : BigInt = (l match {
       case Nil => 0
       case Cons(_, t) => 1 + size(t)
   }) ensuring(res => res >= 0)
 
-  def content(l: List): Set[Int] = l match {
+  def content(l: List): Set[BigInt] = l match {
     case Nil => Set()
     case Cons(i, t) => Set(i) ++ content(t)
   }
 
 
 
-  def insertUnique(l: List, v: Int): List = (l match {
+  def insertUnique(l: List, v: BigInt): List = (l match {
     case Cons(h, t) =>
       if (h == v) {
         l
@@ -37,7 +37,7 @@ object List {
 
 
 
-  def delete(l: List, v: Int): List = {
+  def delete(l: List, v: BigInt): List = {
     choose{ (res: List) =>
       content(res) == content(l) -- Set(v)
     }
