+++ conf/resources/examples/synthesis/01_List_Insert.scala
@@ -4,20 +4,20 @@ import leon.annotation._
 
 object Insert {
   sealed abstract class List
-  case class Cons(head: Int, tail: List) extends List
+  case class Cons(head: BigInt, tail: List) extends List
   case object Nil extends List
 
-  def size(l: List) : Int = (l match {
+  def size(l: List) : BigInt = (l match {
       case Nil => 0
       case Cons(_, t) => 1 + size(t)
   }) ensuring(res => res >= 0)
 
-  def content(l: List): Set[Int] = l match {
-    case Nil => Set.empty[Int]
+  def content(l: List): Set[BigInt] = l match {
+    case Nil => Set.empty[BigInt]
     case Cons(i, t) => Set(i) ++ content(t)
   }
 
-  def insert(in1: List, v: Int) = choose {
+  def insert(in1: List, v: BigInt) = choose {
     (out : List) =>
       content(out) == content(in1) ++ Set(v)
   }
