+++ conf/resources/examples/synthesis/06_List_Split2.scala
@@ -4,16 +4,16 @@ import leon.annotation._
 
 object List {
   sealed abstract class List
-  case class Cons(head: Int, tail: List) extends List
+  case class Cons(head: BigInt, tail: List) extends List
   case object Nil extends List
 
-  def size(l: List) : Int = (l match {
+  def size(l: List) : BigInt = (l match {
       case Nil => 0
       case Cons(_, t) => 1 + size(t)
   }) ensuring(res => res >= 0)
 
-  def content(l: List): Set[Int] = l match {
-    case Nil => Set.empty[Int]
+  def content(l: List): Set[BigInt] = l match {
+    case Nil => Set.empty[BigInt]
     case Cons(i, t) => Set(i) ++ content(t)
   }
 
@@ -24,7 +24,7 @@ object List {
     content(res._1) ++ content(res._2) == content(list) 
   }
 
-  def abs(i : Int) : Int = {
+  def abs(i : BigInt) : BigInt = {
     if(i < 0) -i else i
   } ensuring(_ >= 0)
