+++ conf/resources/examples/tutorials/05_Ex4_InsertionSort.scala
@@ -7,19 +7,19 @@ import leon.Utils._
 
 object InsertionSort {
   sealed abstract class List
-  case class Cons(head:Int,tail:List) extends List
+  case class Cons(head:BigInt,tail:List) extends List
   case class Nil() extends List
 
   sealed abstract class OptInt
-  case class Some(value: Int) extends OptInt
+  case class Some(value: BigInt) extends OptInt
   case class None() extends OptInt
 
-  def size(l : List) : Int = (l match {
+  def size(l : List) : BigInt = (l match {
     case Nil() => 0
     case Cons(_, xs) => 1 + size(xs)
   }) 
 
-  def contents(l: List): Set[Int] = l match {
+  def contents(l: List): Set[BigInt] = l match {
     case Nil() => Set.empty
     case Cons(x,xs) => contents(xs) ++ Set(x)
   }
@@ -40,7 +40,7 @@ object InsertionSort {
 
   /* Inserting element 'e' into a sorted list 'l' produces a sorted list with
    * the expected content and size */
-  def sortedIns(e: Int, l: List): List = {
+  def sortedIns(e: BigInt, l: List): List = {
     require(isSorted(l))
     l match {
       case Nil() => Cons(e,Nil())
