+++ conf/resources/examples/verification/01_Amortized_Queue.scala
@@ -3,19 +3,19 @@ import leon.annotation._
 
 object AmortizedQueue {
   sealed abstract class List
-  case class Cons(head : Int, tail : List) extends List
+  case class Cons(head : BigInt, tail : List) extends List
   case object Nil extends List
 
   sealed abstract class AbsQueue
   case class Queue(front : List, rear : List) extends AbsQueue
 
-  def size(list : List) : Int = (list match {
+  def size(list : List) : BigInt = (list match {
     case Nil => 0
     case Cons(_, xs) => 1 + size(xs)
   }) ensuring(_ >= 0)
 
-  def content(l: List) : Set[Int] = l match {
-    case Nil => Set.empty[Int]
+  def content(l: List) : Set[BigInt] = l match {
+    case Nil => Set.empty[BigInt]
     case Cons(x, xs) => Set(x) ++ content(xs)
   }
   
@@ -49,7 +49,7 @@ object AmortizedQueue {
       Queue(concat(front, reverse(rear)), Nil)
   } ensuring(isAmortized(_))
 
-  def enqueue(queue : AbsQueue, elem : Int) : AbsQueue = (queue match {
+  def enqueue(queue : AbsQueue, elem : BigInt) : AbsQueue = (queue match {
     case Queue(front, rear) => amortizedQueue(front, Cons(elem, rear))
   }) ensuring(isAmortized(_))
 
@@ -60,7 +60,7 @@ object AmortizedQueue {
     }
   } ensuring (isAmortized(_))
 
-  def front(queue : AbsQueue) : Int = {
+  def front(queue : AbsQueue) : BigInt = {
     require(isAmortized(queue) && !isEmpty(queue))
     queue match {
       case Queue(Cons(f, _), _) => f
@@ -68,7 +68,7 @@ object AmortizedQueue {
   }
 
   @induct
-  def propFront(queue : AbsQueue, list : List, elem : Int) : Boolean = {
+  def propFront(queue : AbsQueue, list : List, elem : BigInt) : Boolean = {
     require(!isEmpty(queue) && isAmortized(queue))
     if (asList(queue) == list) {
       list match {
@@ -78,14 +78,14 @@ object AmortizedQueue {
       true
   } holds
 
-  def enqueueAndFront(queue : AbsQueue, elem : Int) : Boolean = {
+  def enqueueAndFront(queue : AbsQueue, elem : BigInt) : Boolean = {
     if (isEmpty(queue))
       front(enqueue(queue, elem)) == elem
     else
       true
   } holds
 
-  def enqueueDequeueThrice(queue : AbsQueue, e1 : Int, e2 : Int, e3 : Int) : Boolean = {
+  def enqueueDequeueThrice(queue : AbsQueue, e1 : BigInt, e2 : BigInt, e3 : BigInt) : Boolean = {
     if (isEmpty(queue)) {
       val q1 = enqueue(queue, e1)
       val q2 = enqueue(q1, e2)
