+++ conf/resources/examples/verification/05_Propositional_Logic.scala
@@ -8,7 +8,7 @@ object PropositionalLogic {
   case class Or(lhs: Formula, rhs: Formula) extends Formula
   case class Implies(lhs: Formula, rhs: Formula) extends Formula
   case class Not(f: Formula) extends Formula
-  case class Literal(id: Int) extends Formula
+  case class Literal(id: BigInt) extends Formula
 
   def simplify(f: Formula): Formula = (f match {
     case And(lhs, rhs) => And(simplify(lhs), simplify(rhs))
@@ -47,7 +47,7 @@ object PropositionalLogic {
     case Literal(_) => true
   }
 
-  def evalLit(id : Int) : Boolean = (id == 42) // could be any function
+  def evalLit(id : BigInt) : Boolean = (id == 42) // could be any function
   def eval(f: Formula) : Boolean = f match {
     case And(lhs, rhs) => eval(lhs) && eval(rhs)
     case Or(lhs, rhs) => eval(lhs) || eval(rhs)
@@ -62,13 +62,13 @@ object PropositionalLogic {
   } holds
 
   // Note that matching is exhaustive due to precondition.
-  def vars(f: Formula): Set[Int] = {
+  def vars(f: Formula): Set[BigInt] = {
     require(isNNF(f))
     f match {
       case And(lhs, rhs) => vars(lhs) ++ vars(rhs)
       case Or(lhs, rhs) => vars(lhs) ++ vars(rhs)
-      case Not(Literal(i)) => Set[Int](i)
-      case Literal(i) => Set[Int](i)
+      case Not(Literal(i)) => Set[BigInt](i)
+      case Literal(i) => Set[BigInt](i)
     }
   }
